# src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: 12 Kingdoms Backend
  version: V2
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /campaigns:
    get:
      tags:
        - campaign-controller-impl
      summary: Retrieve sheet details
      operationId: campaignDetails
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "403":
          description: Forbidden
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignDetailsResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - campaign-controller-impl
      summary: Create a campaign
      operationId: createCampaign
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /auth/signup:
    post:
      tags:
        - auth-controller-impl
      summary: Register a new player
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/signin:
    post:
      tags:
        - auth-controller-impl
      summary: Authenticate existing player
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /campaigns/removeSheets/{campaignId}:
    patch:
      tags:
        - campaign-controller-impl
      summary: Remove sheets from an existing campaign
      operationId: removeSheets
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdListRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /campaigns/addSheets/{campaignId}:
    patch:
      tags:
        - campaign-controller-impl
      summary: Add sheets to an existing campaign
      operationId: addSheets
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdListRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
  /lineages:
    get:
      tags:
        - lineage-controller-impl
      summary: Retrieve player details
      operationId: lineageDetails
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageModel'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /lineages/list:
    get:
      tags:
        - lineage-controller-impl
      summary: List lineages
      operationId: lineageList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdNameListResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /jobs:
    get:
      tags:
        - job-controller-impl
      summary: Retrieve player details
      operationId: jobDetails
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /jobs/list:
    get:
      tags:
        - job-controller-impl
      summary: List lineages
      operationId: jobList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdNameListResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /campaigns/list:
    get:
      tags:
        - campaign-controller-impl
      summary: List campaigns
      operationId: campaignList
      parameters:
        - name: currentPage
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignListResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
  /campaigns/{id}:
    delete:
      tags:
        - campaign-controller-impl
      summary: Delete a campaign
      operationId: deleteCampaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
components:
  schemas:
...
    IdResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
    ErrorDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      properties:
        errorList:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDto'
    CreateCampaignRequest:
      required:
        - name
      type: object
      properties:
        name:
          maxLength: 30
          minLength: 5
          type: string
        notes:
          type: string
        sheetList:
          type: array
          items:
            type: string
            format: uuid
    SignupRequest:
      required:
        - email
        - firstName
        - lastName
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 20
          minLength: 5
          type: string
        password:
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#&()?[{}]:;',?/*~$^+=<>]).{8,20}$"
          type: string
        firstName:
          maxLength: 20
          minLength: 3
          type: string
        lastName:
          maxLength: 30
          minLength: 3
          type: string
        email:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    LoginRequest:
      required:
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 20
          minLength: 5
          type: string
        password:
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#&()?[{}]:;',?/*~$^+=<>]).{8,20}$"
          type: string
    JwtResponse:
      type: object
      properties:
        type:
          type: string
        jwt:
          type: string
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
...
    IdListRequest:
      required:
        - idList
      type: object
      properties:
        idList:
          type: array
          items:
            type: string
            format: uuid
...
    AptitudeModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
...
    JobModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        mainAttribute:
          type: string
          enum:
            - INTELLIGENCE
            - CUNNING
            - CELERITY
            - TENACITY
        physicalPoints:
          type: integer
          format: int32
        mentalPoints:
          type: integer
          format: int32
        physicalPerLevel:
          type: integer
          format: int32
        mentalPerLevel:
          type: integer
          format: int32
        skills:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/SkillModel'
        aptitudes:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AptitudeModel'
    LineageModel:
      required:
        - description
        - language
        - maturity
        - negativeTraitDescription
        - negativeTraitName
        - positiveTraitDescription
        - positiveTraitName
        - size
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        size:
          type: string
        language:
          type: string
        maturity:
          type: string
        positiveTraitName:
          type: string
        positiveTraitDescription:
          type: string
        negativeTraitName:
          type: string
        negativeTraitDescription:
          type: string
...
    SkillModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        cost:
          type: integer
          format: int32
        energyType:
          type: string
          enum:
            - PHYSICAL
            - MENTAL
        skillType:
          type: string
          enum:
            - INTERPRETATION
            - COMBAT
            - BASIC
        range:
          type: integer
          format: int32
        skillLevel:
          type: integer
          format: int32
...
    IdNameDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    IdNameListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/IdNameDto'
    CampaignDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        notes:
          type: string
        sheets:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/CampaignSheetDto'
    CampaignSheetDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        level:
          type: integer
          format: int32
        mentalCurrent:
          type: integer
          format: int32
        mentalTotal:
          type: integer
          format: int32
        physicalCurrent:
          type: integer
          format: int32
        physicalTotal:
          type: integer
          format: int32
        lineage:
          type: string
        background:
          type: string
    BasicCampaignDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    CampaignListResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        currentPage:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        list:
          type: array
          items:
            $ref: '#/components/schemas/BasicCampaignDto'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT